More changes to perform all tests
