+++ cornflow-server/cornflow/endpoints/__init__.py
@@ -15,6 +15,7 @@ from .case import (
     CaseToInstance,
     CaseCompare,
 )
+
 from .dag import (
     DAGDetailEndpoint,
     DAGEndpointManual,
@@ -31,7 +32,6 @@ from .execution import (
 )
 
 from .health import HealthEndpoint
-from .token import TokenEndpoint
 
 from .instance import (
     InstanceEndpoint,
@@ -41,18 +41,14 @@ from .instance import (
 )
 
 from .login import LoginOpenAuthEndpoint, LoginEndpoint
-
 from .permission import PermissionsViewRoleEndpoint, PermissionsViewRoleDetailEndpoint
-from .roles import (
-    RolesListEndpoint,
-    RoleDetailEndpoint,
-    UserRoleListEndpoint,
-    UserRoleDetailEndpoint,
-)
-from .schemas import SchemaDetailsEndpoint, SchemaEndpoint
-from .signup import SignUpEndpoint
-from .user import UserEndpoint, UserDetailsEndpoint, ToggleUserAdmin, RecoverPassword
 
+from .roles import RolesListEndpoint, RoleDetailEndpoint
+
+from .schemas import SchemaDetailsEndpoint, SchemaEndpoint
+from .token import TokenEndpoint
+from .user import UserEndpoint, UserDetailsEndpoint, ToggleUserAdmin, RecoverPassword
+from .user_role import UserRoleListEndpoint, UserRoleDetailEndpoint
 
 resources = [
     dict(resource=InstanceEndpoint, urls="/instance/", endpoint="instance"),
@@ -109,7 +105,6 @@ resources = [
         urls="/user/<int:user_id>/<int:make_admin>/",
         endpoint="user-admin",
     ),
-    dict(resource=SignUpEndpoint, urls="/signup/", endpoint="signup"),
     dict(resource=TokenEndpoint, urls="/token/", endpoint="token"),
     dict(resource=SchemaEndpoint, urls="/schema/", endpoint="schema"),
     dict(
