+++ cornflow-server/cornflow/tests/unit/test_users.py
@@ -26,7 +26,6 @@ from cornflow.tests.const import (
     LOGIN_URL,
     SIGNUP_URL,
     USER_URL,
-    RECOVER_PASSWORD_URL,
 )
